# openapi.yaml
openapi: 3.0.0
info:
  title: Dokany API
  description: منصة دكاني - واجهات برمجة التطبيقات
  version: 1.1.0
  contact:
    name: Dokany Team
    url: https://dokany.com
servers:
  - url: https://api.dokany.com
    description: Production Server
  - url: http://localhost:3000
    description: Development Server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    storeIdHeader:
      type: apiKey
      in: header
      name: x-store-id
  schemas:
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        code:
          type: string
          example: "VALIDATION_ERROR"
        message:
          type: string
          example: "بيانات غير صالحة"
...
    Category:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        image:
          type: string
        parentId:
          type: string
        isActive:
          type: boolean
        order:
          type: integer
...
security:
  - bearerAuth: []
  - storeIdHeader: []
paths:
...
  /store/{slug}/categories:
    get:
      tags: [Categories]
      summary: جلب قائمة التصنيفات
      x-implemented: true
      operationId: getCategories
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: قائمة التصنيفات
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Category'

    post:
      tags: [Categories]
      summary: إنشاء تصنيف جديد
      x-implemented: true
      security:
        - bearerAuth: []
        - storeIdHeader: []
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                description: { type: string }
                parentId: { type: string }
                order: { type: integer }
                isActive: { type: boolean }
      responses:
        '201':
          description: تم الإنشاء
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Category'
        '400':
          description: بيانات غير صالحة
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: غير مصرح
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /store/{slug}/categories/{categoryId}:
    get:
      tags: [Categories]
      summary: جلب تفاصيل تصنيف
      x-implemented: false
      description: |
        ⚠️ **غير منفذ حالياً**
        يجب إضافة هذا الـ Endpoint في `workers/routes/categories.ts`.
        يعيد تفاصيل التصنيف مع قائمة المنتجات التابعة له.
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
        - name: categoryId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: تفاصيل التصنيف
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Category'
        '404':
          description: التصنيف غير موجود
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags: [Categories]
      summary: تحديث تصنيف
      x-implemented: true
      security:
        - bearerAuth: []
        - storeIdHeader: []
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
        - name: categoryId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                description: { type: string }
                parentId: { type: string }
                order: { type: integer }
                isActive: { type: boolean }
      responses:
        '200':
          description: تم التحديث
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Category'

    delete:
      tags: [Categories]
      summary: حذف تصنيف
      x-implemented: true
      security:
        - bearerAuth: []
        - storeIdHeader: []
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
        - name: categoryId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: تم الحذف
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      message:
                        type: string

  # ============================================================
  # 📦 PRODUCTS
  # ============================================================
...